Remove SharedKafkaProxyContext
