+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/ClusterService.java
@@ -53,7 +53,7 @@ public class ClusterService
         ProxyModelBuilder proxyModelBuilder = ProxyModelBuilder.contextBuilder(context);
         ProxyModel model = proxyModelBuilder.build(primary, context);
         Stream<Service> serviceStream = model.clustersWithValidIngresses().stream()
-                .filter(cluster -> !SharedKafkaProxyContext.isBroken(context, cluster))
+                .filter(cluster -> !KafkaProxyReconciler.isBroken(context, cluster))
                 .flatMap(cluster -> model.ingressModel().clusterIngressModel(cluster).map(ProxyIngressModel.VirtualClusterIngressModel::services).orElse(Stream.empty()));
         return serviceStream.collect(toByNameMap());
     }
