+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/resolver/ClusterConditionUnresolvedDependencyReporter.java
@@ -15,7 +15,7 @@ import io.kroxylicious.kubernetes.api.common.LocalRef;
 import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxy;
 import io.kroxylicious.kubernetes.api.v1alpha1.VirtualKafkaCluster;
 import io.kroxylicious.kubernetes.operator.ClusterCondition;
-import io.kroxylicious.kubernetes.operator.SharedKafkaProxyContext;
+import io.kroxylicious.kubernetes.operator.KafkaProxyReconciler;
 
 import static io.kroxylicious.kubernetes.operator.ResourcesUtil.name;
 
@@ -32,7 +32,7 @@ public record ClusterConditionUnresolvedDependencyReporter(Context<KafkaProxy> c
         LocalRef<?> firstUnresolvedDependency = unresolvedDependencies.stream()
                 .sorted(comparator).findFirst()
                 .orElseThrow();
-        SharedKafkaProxyContext.addClusterCondition(context, cluster,
+        KafkaProxyReconciler.addClusterCondition(context, cluster,
                 ClusterCondition.refNotFound(name(cluster), firstUnresolvedDependency));
     }
 }
